pyproject.toml
[project]
name = "equity-pilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "mcp>=1.27.0",
    "ollama>=0.6.2",
    "python-dotenv>=1.0.0",
    "yfinance>=1.3.0",
]
